+++ lib/fcode_bloc.dart
@@ -1,4 +1,4 @@
-// Copyright 2019 The Fcode Labs Authors. All rights reserved.
+// Copyright 2020 The Fcode Labs Authors. All rights reserved.
 // Use of this source code is governed by a MIT license that can be
 // found in the LICENSE file.
 
@@ -7,6 +7,9 @@ library fcode_bloc;
 export 'src/db/cloud_support.dart';
 export 'src/db/model/db_model.dart';
 export 'src/db/model/db_model_i.dart';
+export 'src/db/model/serializer/document_reference_serializer.dart';
+export 'src/db/model/serializer/geo_point_serializer.dart';
+export 'src/db/model/serializer/timestamp_serializer.dart';
 export 'src/db/reference_handler.dart';
 export 'src/db/references_handler.dart';
 export 'src/db/repo/firebase_repository.dart';
@@ -16,5 +19,4 @@ export 'src/db/spec/firebase/complex_specification.dart';
 export 'src/db/spec/firebase/firebase_specification.dart';
 export 'src/db/spec/firebase/text_start_with_specification.dart';
 export 'src/db/spec/specification.dart';
-export 'src/log/log.dart';
 export 'src/util/algo.dart';